crawler-service/requirements.txt
functions-framework~=3.5.0
pydantic~=2.11.8
langchain-google-vertexai~=2.1.0
google-cloud-firestore~=2.21.0
google-cloud-aiplatform~=1.115.0
aiohttp~=3.12.15
beautifulsoup4~=4.13.5
tenacity~=9.1.2